search-app/pyproject.toml
[project]
name = "enterprise-searchapp"
version = "0.1.0"
description = "Enterprise-grade Gradio + FastAPI search app with OCI PostgreSQL + pgvector backend (RAG, semantic, hybrid, full-text)"
authors = [{ name = "Your Org", email = "devnull@example.com" }]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "Apache-2.0" }

# Core runtime dependencies
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "gradio>=4.41.0,<5",
  "pydantic>=2.6.0",
  "python-multipart>=0.0.9",
  "psycopg[binary,pool]>=3.2.0",

  "beautifulsoup4>=4.12.2",
  "pypdf>=4.2.0",
  "sentence-transformers>=3.0.0",
  "tqdm>=4.66.0",
  "python-dotenv>=1.0.1",
  "oci>=2.129.0",
  "python-docx>=1.1.0",
]

[project.optional-dependencies]
# Optional LLM providers
openai = ["openai>=1.60.0"]

[project.urls]
Homepage = "https://github.com/your-org/enterprise-searchapp"

[project.scripts]
searchapp = "app.main:main"

[tool.uv]
# Ensure uv respects this project file for sync/install
package = true

[tool.ruff]
line-length = 120

[tool.pyright]
reportMissingImports = "warning"
